+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Updated minimum required Okteto CLI version to 3.0.0
 - Remove "okteto create" command since the CLI doesn't support it anymore
 - Removed support for Okteto Stacks or Okteto Manifest 1.0 since they have been fully deprecated
+- Namespace is no longer read from the manifest; instead, you need to set it via the `context` command
 - Updated dependencies
 
 ## 0.4.15